Added multiple concurrent DescriptionTokenizer threads.
