+++ 100_bank_transaction_descriptions.csv
@@ -1,0 +1,100 @@
+"CHECKCARD 0120 24HOUR FITNESS USA,INC 800-432-6348 CA..."
+"CHECKCARD 0221 24HOUR FITNESS USA,INC 800-432-6348 CA..."
+"CHECKCARD 0321 24HOUR FITNESS USA,INC 800-432-6348 CA..."
+"CHECKCARD 0420 24HOUR FITNESS USA,INC 800-432-6348 CA..."
+"CHECKCARD 0426 24HOUR FITNESS USA,INC 800-432-6348 CA..."
+"CHECKCARD 0520 24HOUR FITNESS USA,INC 800-432-6348 CA..."
+"CHECKCARD 0621 24HOUR FITNESS USA,INC 800-432-6348 CA..."
+"CHECKCARD 0720 24HOUR FITNESS USA,INC 800-432-6348 CA..."
+"CHECKCARD 0820 24HOUR FITNESS USA,INC 800-432-6348 CA..."
+"CHECKCARD 0920 24HOUR FITNESS USA,INC 800-432-6348 CA..."
+"CHECKCARD 1020 24HOUR FITNESS USA,INC 800-432-6348 CA..."
+"CHECKCARD 1120 24HOUR FITNESS USA,INC 800-432-6348 CA..."
+"CHECKCARD 1220 24HOUR FITNESS USA,INC 800-432-6348 CA..."
+"CHECKCARD 1227 BCF-COQUITLAM,QUEEN OF VICTORIA BC 74500012362461680275931"
+"CHECKCARD 1227 BCF-COQUITLAM,QUEEN OF VICTORIA BC 74500012362461680275931..."
+"CONCUR DES:EXPENSE ID:C008N1RUX000 INDN:Key, Joe A. CO ID:1911608052 PPD..."
+"CONCUR DES:EXPENSE ID:C00A9UJCX006 INDN:Key, Joe A. CO ID:1911608052 PPD..."
+"H S B C 04/12 #000887347 WITHDRWL H S B C / JOSE AZUETA, FEE"
+"H S B C 04/12 #000887347 WITHDRWL H S B C / JOSE AZUETA, INTERNATIONAL..."
+"H S B C 04/12 #000887347 WITHDRWL H S B C / JOSE AZUETA,"
+"YODLEE.COM INC DES:DIRECT DEP ID:315018338477QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:475019277469QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:515046671679QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:516042312983QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:516042312984QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:517046621180QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:519046304065QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:569025025433QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:575027196035QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:588025780457QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:590017342045QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:612026847301QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:625043170387QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:636041888967QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:640025839973QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:640025839974QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:654038835037QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:669033410119QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:681023381652QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:689051101128QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:696048239467QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:703030941426QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:714044355791QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:715043619496QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:740017497838QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:740017497839QBC INDN:KEY,JOE A CO..."
+"YODLEE.COM INC DES:DIRECT DEP ID:793028169125QBC INDN:KEY,JOE A CO..."
+*SCOTIABANK IN 04/07 #000684996 WITHDRWL SCOTIABANK IN/ BCO INVERLAT
+*SCOTIABANK IN 04/07 #000685080 WITHDRWL SCOTIABANK IN/ BCO INVERLAT
+*SCOTIABANK IN 04/08 #000704407 WITHDRWL SCOTIABANK IN/ BCO INVERLAT
+7-ELEVEN        77 NEWELL RD           EAST PALO ALTCAUS
+7-ELEVEN 02/03 #000186601 PURCHASE 1430 FRUITDALE AV SAN JOSE CA
+7-ELEVEN 03/10 #000101978 PURCHASE 1430 FRUITDALE AV SAN JOSE CA
+76                  MENLO PARK      CAUS
+A & S 76        710 WILLOW RD          MENLO PARK   CAUS
+AM VALERO FOOD 08/16 #000099854 PURCHASE 1502 E YOSEMITE A MADERA CA
+ARAMARK SHORELINE A MOUNTAIN VIEW   CAUS
+ARCO PAYPOINT 01/22 #000490226 PURCHASE 470 RALSTON AVE BELMONT CA
+ARCO PAYPOINT 01/31 #000311596 PURCHASE 470 RALSTON AVE BELMONT CA
+ARCO PAYPOINT 02/16 #000888874 PURCHASE 1040 SUNNYVALE - SUNNYVALE CA
+ARCO PAYPOINT 04/05 #000079544 PURCHASE 1040 SUNNYVALE - SUNNYVALE CA
+ARCO PAYPOINT 05/31 #000753861 PURCHASE 1040 SUNNYVALE - SUNNYVALE CA
+ARCO PAYPOINT 06/08 #000431971 PURCHASE 18605 MONTEREY RD MORGAN HILL CA
+ARCO PAYPOINT 06/08 #000915386 PURCHASE 18605 MONTEREY RD MORGAN HILL CA
+ARCO PAYPOINT 08/22 #000589150 PURCHASE 1040 SUNNYVALE - SUNNYVALE CA
+ARCO PAYPOINT 09/03 #000802492 PURCHASE 470 RALSTON AVE BELMONT CA
+ARCO PAYPOINT 10/02 #000577272 PURCHASE 1040 SUNNYVALE - SUNNYVALE CA
+ARCO PAYPOINT 11/27 #000503476 PURCHASE ARCO PAYPOINT SUNNYVALE CA
+ATM Withdrawal BENDERS BAR     806 S VAN NESS AVE     SAN FRAN     CAUS
+BALOMPIE CAFE RESTA SAN FRANCISCO   CAUS
+BART 16TH ST MI 2000 MISSION STREET    SAN FRANCISCOCAUS
+BART MILLBRAE STAT  5104646979      CAUS
+BART SAN BRUNO 01/21 #000149366 PURCHASE BART SAN BRUNO ST SAN BRUNO CA
+BART SAN BRUNO 06/12 #000153033 PURCHASE BART SAN BRUNO ST SAN BRUNO CA
+BART SAN BRUNO 06/19 #000406956 PURCHASE BART SAN BRUNO ST SAN BRUNO CA
+BART SAN BRUNO 08/01 #000266625 PURCHASE BART SAN BRUNO ST SAN BRUNO CA
+BART SAN BRUNO 08/13 #000190509 PURCHASE BART SAN BRUNO ST SAN BRUNO CA
+BART SAN BRUNO 12/02 #000007164 PURCHASE BART SAN BRUNO ST SAN BRUNO CA
+BCF - 1600 SAR 06/02 #000065900 PURCHASE BCF - 1600 SARATO SAN JOSE CA
+BEVERAGES & MO 01/31 #000570224 PURCHASE 1247 W EL CAMINO SUNNYVALE CA
+BEVERAGES & MO 02/22 #000609042 PURCHASE 1247 W EL CAMINO SUNNYVALE CA
+BEVERAGES & MO 03/01 #000646355 PURCHASE 1247 W EL CAMINO SUNNYVALE CA
+BEVERAGES & MO 03/16 #000586110 PURCHASE 1247 W EL CAMINO SUNNYVALE CA
+BEVERAGES & MO 09/05 #000640591 PURCHASE 1247 W EL CAMINO SUNNYVALE CA
+BEVERAGES & MO 11/07 #000886296 PURCHASE BEVERAGES & MOR SUNNYVALE CA
+BIG 5 SPORTING 05/06 #000024778 PURCHASE BIG 5 SPORTING GO SUNNYVALE CA
+BIG 5 SPORTING 06/18 #000014839 PURCHASE BIG 5 SPORTING GO SUNNYVALE CA
+BIG 5 SPTG GDS 02/20 #000462508 PURCHASE BIG 5 SPTG GDS-00 SUNNYVALE CA
+BKOFAMERICA ATM 01/18 #000001041 WITHDRWL SUNNYVALE MAIN SUNNYVALE CA
+BKOFAMERICA ATM 04/07 #000008734 WITHDRWL LAX-TERMINAL 5 D LOS ANGELES...
+BKOFAMERICA ATM 05/05 #000007074 WITHDRWL SUNNYVALE MAIN SUNNYVALE CA
+BKOFAMERICA ATM 09/21 #000001290 WITHDRWL SUNNYVALE MAIN SUNNYVALE CA
+BNP 10/24 #000766600 WITHDRWL >49740AGENCE P PARIS 09
+BROOKSHIRE #13 06/27 #000554402 PURCHASE 7713 HWY 165 COLUMBIA LA
+CA TLR cash withdrawal from CHK 2290 Banking Ctr SUNNYVALE INDUSTRIAL...
+CA TLR cash withdrawal from CHK 2290 Banking Ctr SUNNYVALE MAIN OFFICE...
+CAFE ZOE            MENLO PARK      CAUS
+CENTURY THEATRES 42 UNION CITY      CAUS
+CENTURY THEATRES 48 REDWOOD CITY    CAUS
+CHASE 06/27 #000530159 WITHDRWL 2600 CYPRESS ST W MONROE LA
