+++ config_3.json
@@ -1,0 +1,20 @@
+{
+	"concurrency" : 10,
+	"input" : {
+		"filename" : "../100_bank_transaction_descriptions.csv",
+		"encoding" : "utf-8"
+	},
+	"logging" : {
+		"level" : "info",
+		"path" : "../logs.log",
+		"formatter" : "%(asctime)s - %(name)s - %(levelname)s - %(message)s",
+		"console" : true
+	},
+	"output" : {
+		"results" : {
+			"fields" : [ "BUSINESSSTANDARDNAME", "HOUSE", "PREDIR", "STREET",
+				"STRTYPE", "CITYNAME", "STATE", "ZIP", "pin.location"
+			]
+		}
+	}
+}
